pyproject.toml
[tool.poetry]
name = "bus-routes"
version = "0.1.0"
description = ""
authors = ["Thomas Richards <thomas.c.richards@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
